Change CreatedAt type to long long

+++ secrethub_wrapper.go
@@ -4,7 +4,7 @@ package main
 struct Secret {
 	int Version;
 	char* Data;
-	long CreatedAt;
+	long long CreatedAt;
 };
 #include <stdbool.h>
 */
@@ -32,7 +32,7 @@ func Read(path *C.char, errMessage **C.char) C.struct_Secret {
 	return C.struct_Secret{
 		Version: C.int(secret.Version),
 		Data:    C.CString(string(secret.Data)),
-		CreatedAt: C.long(secret.CreatedAt.Unix()),
+		CreatedAt: C.longlong(secret.CreatedAt.Unix()),
 	}
 }
